requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
Pillow==9.0.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pytz==2021.3
sqlparse==0.4.2
